backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
mysqlclient>=2.2.0
anthropic>=0.40.0
weasyprint>=62.0
